Added ansible configuration to setup bastion host for Openshift installation.
The ansible inventory file is generated and reloaded from the ansible playbook, this is done from the play that runs on the localhost, oherwise the bastion group does not exist.
A jinja2 template is used to create the install-config.yaml configuration file

The region_name terraform variable uses short region names, which are compatible with what is used in the install-config.yaml configuration file.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>