Added code to deploy a private cluster not accessible from the Internet
The terraform variable cluster_scope defines if the cluster will be public or private.  This var contains a validation block to make sure that only one of the two valid values are assigned to it.

A module to conditionally create the nat gateway infraestrucutre is added to terraform.  The creation or not depends on the value of the terraform variable outbound_type, if it is LoadBalancer the nat gateway is not created, otherwise it is created.

The install-config.yaml jinja template contains logic to define the azure variables depending on the output variables from terraform.  The subnet names are extracted with a regular expression, and because the result is a list of one member, only the first element is selected to get a string.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>